pyproject.toml
[tool.poetry]
name = "sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["nakatuba <nakatuba0626@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.4.2"
mecab-python3 = "^1.0.3"
pandas = "^1.2.4"
scikit-learn = "^0.24.2"
torch = "^1.8.1"
torchtext = "^0.10.0"
wandb = "^0.10.33"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
flake8 = "^3.9.2"
isort = "^5.8.0"
jupyterlab = "^3.0.16"
jupyterlab-code-formatter = "^1.4.10"
jupyterlab-kite = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
